refactor: 💡 rename typs -> types and change import
