+++ src/pages/contact/index.page.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useContactHandler } from '@/pages/contact/hooks/useContactHandler';
 import { ContactForm } from '@/pages/contact/components/ContactForm';
 import { loadSessionData } from '@/pages/contact/confirm/logics/loadSession';
-import { UserInput } from '@/pages/contact/type';
+import { UserInput } from './types';
 
 export default function Contact() {
   const {
